Add the ArrayUtils::explodeTrim() method
